refactor: relocate ConnectionService to a new directory
